+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 * Infrastructure changes
 * Include .vscode folder
 * Remove interactive mode
+* Add support for solution level content in templates (`_content`)
 
 ### 1.4.2 - 2017-06-03
 * Add support for `FORGE_TEMPLATE_DIR` environment variable